# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "工業數據採集閘道系統 API 文檔",
        "title": "Go Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/datalink/studio-v2/workspace/recording-start": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Studio V2 Write Groups"
                ],
                "summary": "Start recording for selected devices and groups",
                "parameters": [
                    {
                        "description": "Scoped start identity and expected revisions",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/workspace.RecordingStartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RecordingStartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/datalink/studio-v2/workspace/recording-start/operations/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Studio V2 Write Groups"
                ],
                "summary": "Read scoped recording start progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RecordingStartResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/datalink/studio-v2/workspace/write-groups/basic/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Studio V2 Write Groups"
                ],
                "summary": "Save or reuse one device's basic managed group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Basic managed draft and expected revisions",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.WriteGroupMutationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.WriteGroupSaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handlers.TypedAPIErrorEnvelope"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.RecordingStartResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/workspace.RecordingStartOperation"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.TypedAPIErrorEnvelope": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "dirty_state": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "operation_id": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "retryable": {
                    "type": "boolean"
                },
                "settings_revision": {
                    "type": "string"
                },
                "workspace_revision": {
                    "type": "string"
                }
            }
        },
...
        "handlers.WriteGroupMutationRequest": {
            "type": "object",
            "properties": {
                "expected_connector_revision": {
                    "type": "string"
                },
                "expected_group_revision": {
                    "type": "string"
                },
                "expected_workspace_revision": {
                    "type": "string"
                },
                "group": {
                    "$ref": "#/definitions/workspace.WriteGroup"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.WriteGroupSaveResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/workspace.WriteGroupSaveResult"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "recordingplan.SchemaOperationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "succeeded",
                "partial",
                "failed",
                "unknown"
            ],
            "x-enum-varnames": [
                "SchemaOperationPending",
                "SchemaOperationRunning",
                "SchemaOperationSucceeded",
                "SchemaOperationPartial",
                "SchemaOperationFailed",
                "SchemaOperationUnknown"
            ]
        },
...
        "workspace.RecordingStartDeviceProgress": {
            "type": "object",
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "device_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "workspace.RecordingStartGroupIntent": {
            "type": "object",
            "properties": {
                "draft": {
                    "$ref": "#/definitions/workspace.WriteGroup"
                },
                "expected_connector_revision": {
                    "type": "string"
                },
                "expected_group_revision": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                }
            }
        },
        "workspace.RecordingStartGroupProgress": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "boolean"
                },
                "applied_revision": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "group_revision": {
                    "type": "string"
                },
                "ready": {
                    "type": "boolean"
                },
                "saved": {
                    "type": "boolean"
                }
            }
        },
        "workspace.RecordingStartOperation": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspace.RecordingStartDeviceProgress"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspace.RecordingStartGroupProgress"
                    }
                },
                "intent_digest": {
                    "type": "string"
                },
                "next_action": {
                    "type": "string"
                },
                "operation_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "setup_revision": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/recordingplan.SchemaOperationStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
        "workspace.RecordingStartRequest": {
            "type": "object",
            "properties": {
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expected_workspace_revision": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspace.RecordingStartGroupIntent"
                    }
                },
                "readiness_token": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "settings_revision": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                },
                "workspace_revision": {
                    "type": "string"
                }
            }
        },
        "workspace.WriteGroup": {
            "type": "object",
            "properties": {
                "applied_revision": {
                    "type": "string"
                },
                "basic_managed_device_id": {
                    "description": "BasicManagedDeviceID is read from the create-once key, never caller-assigned.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "destination": {
                    "$ref": "#/definitions/workspace.WriteGroupDestination"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspace.WriteGroupMember"
                    }
                },
                "migration": {
                    "$ref": "#/definitions/workspace.WriteGroupMigration"
                },
                "name": {
                    "type": "string"
                },
                "revision": {
                    "type": "string"
                },
                "row_policy": {
                    "$ref": "#/definitions/workspace.WriteGroupRowPolicy"
                },
                "status": {
                    "$ref": "#/definitions/workspace.WriteGroupStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                },
                "write_policy": {
                    "$ref": "#/definitions/workspace.WriteGroupWritePolicy"
                }
            }
        },
        "workspace.WriteGroupDestination": {
            "type": "object",
            "properties": {
                "connector_id": {
                    "type": "string"
                },
                "connector_revision": {
                    "type": "string"
                },
                "database": {
                    "type": "string"
                },
                "schema_digest": {
                    "type": "string"
                },
                "schema_revision": {
                    "type": "string"
                },
                "storage_strategy": {
                    "$ref": "#/definitions/workspace.WriteGroupStorageStrategy"
                },
                "table_name": {
                    "type": "string"
                },
                "table_schema": {
                    "type": "string"
                }
            }
        },
...
        "workspace.WriteGroupMember": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "entity_key": {
                    "description": "EntityKey is the raw legacy group key used only to partition buffered\nrows. It is deliberately not interpreted as a destination SQL value.",
                    "type": "string"
                },
                "mapping_revision": {
                    "type": "string"
                },
                "max_age_seconds": {
                    "type": "integer"
                },
                "measurement_id": {
                    "type": "string"
                },
                "point_id": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "source_revision": {
                    "type": "string"
                },
                "tag_id": {
                    "type": "string"
                },
                "target_column": {
                    "type": "string"
                }
            }
        },
        "workspace.WriteGroupMigration": {
            "type": "object",
            "properties": {
                "adapter_version": {
                    "type": "string"
                },
                "legacy_row_group_id": {
                    "type": "string"
                },
                "review_result": {
                    "type": "string"
                },
                "source_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source_kind": {
                    "type": "string"
                },
                "source_revision": {
                    "type": "string"
                },
                "target_mapping_points": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "workspace.WriteGroupRowPolicy": {
            "type": "object",
            "properties": {
                "allowed_lateness_seconds": {
                    "type": "integer"
                },
                "bucket_start_column": {
                    "type": "string"
                },
                "device_id_column": {
                    "type": "string"
                },
                "entity_key_column": {
                    "type": "string"
                },
                "group_id_column": {
                    "type": "string"
                },
                "group_key_columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "incomplete_policy": {
                    "type": "string"
                },
                "interval_seconds": {
                    "type": "integer"
                },
                "provenance_column": {
                    "type": "string"
                },
                "quality_column": {
                    "type": "string"
                },
                "record_key_column": {
                    "type": "string"
                },
                "unique_key_columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "value_column": {
                    "type": "string"
                }
            }
        },
        "workspace.WriteGroupSaveResult": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/workspace.WriteGroup"
                },
                "workspace_revision": {
                    "type": "string"
                }
            }
        },
        "workspace.WriteGroupStatus": {
            "type": "string",
            "enum": [
                "draft",
                "ready",
                "running",
                "disabled",
                "deleted"
            ],
            "x-enum-varnames": [
                "WriteGroupStatusDraft",
                "WriteGroupStatusReady",
                "WriteGroupStatusRunning",
                "WriteGroupStatusDisabled",
                "WriteGroupStatusDeleted"
            ]
        },
        "workspace.WriteGroupStorageStrategy": {
            "type": "string",
            "enum": [
                "managed",
                "custom"
            ],
            "x-enum-varnames": [
                "WriteGroupStorageStrategyManaged",
                "WriteGroupStorageStrategyCustom"
            ]
        },
        "workspace.WriteGroupWritePolicy": {
            "type": "object",
            "properties": {
                "dedupe_capability": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                }
            }
        }
    }
}